Hack for routeRequest for interim cluster performance
Added some "fast" preprocessing to counters to deduce that there are
zones without enough nodes/partitions to meet intended storage def
requirements. Fast is in quotes because it is an O(# nodes)
pre-processing step. In general, this is much smaller than O(#
partitions). The code is ugly and fragile. It smells. I'd prefer to go
with "slow" performance while interim cluster is deployed.

Committing the ProfileRouteRequest tool from Vinoth to share with
folks. Note that path is hard coded in this tool at this time.