Hack to scope rebalance controller/scheduler hardening
Many TODOs to document areas of exception throwing/catching that are
weird and how to handle killing rebalance controller by hand.

Changed some try/catch stuff in the hopes of actually rolling back
upon task failure or rebalance controller kill.

Change behavior to drain the queue of tasks to be scheduled as soon as
one of them fails.